+++ lib/src/page.dart
@@ -23,6 +23,7 @@ import 'pipe.dart';
 part 'disposer.dart';
 part 'page_logic.dart';
 part 'page_object.dart';
+part 'view_logic.dart';
 
 abstract class FPage<T extends FPageLogic> {
   final _po = _FPageObject<T>._();
@@ -186,8 +187,7 @@ class _PagePipe {
     return StreamBuilder5<T1, T2, T3, T4, T5>(
       streams: StreamTuple5(pipe1.stream, pipe2.stream, pipe3.stream, pipe4.stream, pipe5.stream),
       initialData: StreamInitialDataTuple5(pipe1.value, pipe2.value, pipe3.value, pipe4.value, pipe5.value),
-      builder: (context, s1, s2, s3, s4, s5) =>
-          listener(s1.data as T1, s2.data as T2, s3.data as T3, s4.data as T4, s5.data as T5),
+      builder: (context, s1, s2, s3, s4, s5) => listener(s1.data as T1, s2.data as T2, s3.data as T3, s4.data as T4, s5.data as T5),
     );
   }
 
@@ -202,10 +202,8 @@ class _PagePipe {
   }) {
     return StreamBuilder6<T1, T2, T3, T4, T5, T6>(
       streams: StreamTuple6(pipe1.stream, pipe2.stream, pipe3.stream, pipe4.stream, pipe5.stream, pipe6.stream),
-      initialData:
-          StreamInitialDataTuple6(pipe1.value, pipe2.value, pipe3.value, pipe4.value, pipe5.value, pipe6.value),
-      builder: (context, s1, s2, s3, s4, s5, s6) =>
-          listener(s1.data as T1, s2.data as T2, s3.data as T3, s4.data as T4, s5.data as T5, s6.data as T6),
+      initialData: StreamInitialDataTuple6(pipe1.value, pipe2.value, pipe3.value, pipe4.value, pipe5.value, pipe6.value),
+      builder: (context, s1, s2, s3, s4, s5, s6) => listener(s1.data as T1, s2.data as T2, s3.data as T3, s4.data as T4, s5.data as T5, s6.data as T6),
     );
   }
 
@@ -220,12 +218,9 @@ class _PagePipe {
     required Widget Function(T1 val1, T2 val2, T3 val3, T4 val4, T5 val5, T6 val6, T7 val7) listener,
   }) {
     return StreamBuilder7<T1, T2, T3, T4, T5, T6, T7>(
-      streams: StreamTuple7(
-          pipe1.stream, pipe2.stream, pipe3.stream, pipe4.stream, pipe5.stream, pipe6.stream, pipe7.stream),
-      initialData: StreamInitialDataTuple7(
-          pipe1.value, pipe2.value, pipe3.value, pipe4.value, pipe5.value, pipe6.value, pipe7.value),
-      builder: (context, s1, s2, s3, s4, s5, s6, s7) => listener(
-          s1.data as T1, s2.data as T2, s3.data as T3, s4.data as T4, s5.data as T5, s6.data as T6, s7.data as T7),
+      streams: StreamTuple7(pipe1.stream, pipe2.stream, pipe3.stream, pipe4.stream, pipe5.stream, pipe6.stream, pipe7.stream),
+      initialData: StreamInitialDataTuple7(pipe1.value, pipe2.value, pipe3.value, pipe4.value, pipe5.value, pipe6.value, pipe7.value),
+      builder: (context, s1, s2, s3, s4, s5, s6, s7) => listener(s1.data as T1, s2.data as T2, s3.data as T3, s4.data as T4, s5.data as T5, s6.data as T6, s7.data as T7),
     );
   }
 
@@ -241,12 +236,10 @@ class _PagePipe {
     required Widget Function(T1 val1, T2 val2, T3 val3, T4 val4, T5 val5, T6 val6, T7 val7, T8 val8) listener,
   }) {
     return StreamBuilder8<T1, T2, T3, T4, T5, T6, T7, T8>(
-      streams: StreamTuple8(pipe1.stream, pipe2.stream, pipe3.stream, pipe4.stream, pipe5.stream, pipe6.stream,
-          pipe7.stream, pipe8.stream),
-      initialData: StreamInitialDataTuple8(
-          pipe1.value, pipe2.value, pipe3.value, pipe4.value, pipe5.value, pipe6.value, pipe7.value, pipe8.value),
-      builder: (context, s1, s2, s3, s4, s5, s6, s7, s8) => listener(s1.data as T1, s2.data as T2, s3.data as T3,
-          s4.data as T4, s5.data as T5, s6.data as T6, s7.data as T7, s8.data as T8),
+      streams: StreamTuple8(pipe1.stream, pipe2.stream, pipe3.stream, pipe4.stream, pipe5.stream, pipe6.stream, pipe7.stream, pipe8.stream),
+      initialData: StreamInitialDataTuple8(pipe1.value, pipe2.value, pipe3.value, pipe4.value, pipe5.value, pipe6.value, pipe7.value, pipe8.value),
+      builder: (context, s1, s2, s3, s4, s5, s6, s7, s8) =>
+          listener(s1.data as T1, s2.data as T2, s3.data as T3, s4.data as T4, s5.data as T5, s6.data as T6, s7.data as T7, s8.data as T8),
     );
   }
 
@@ -263,12 +256,10 @@ class _PagePipe {
     required Widget Function(T1 val1, T2 val2, T3 val3, T4 val4, T5 val5, T6 val6, T7 val7, T8 val8, T9 val9) listener,
   }) {
     return StreamBuilder9<T1, T2, T3, T4, T5, T6, T7, T8, T9>(
-      streams: StreamTuple9(pipe1.stream, pipe2.stream, pipe3.stream, pipe4.stream, pipe5.stream, pipe6.stream,
-          pipe7.stream, pipe8.stream, pipe9.stream),
-      initialData: StreamInitialDataTuple9(pipe1.value, pipe2.value, pipe3.value, pipe4.value, pipe5.value, pipe6.value,
-          pipe7.value, pipe8.value, pipe9.value),
-      builder: (context, s1, s2, s3, s4, s5, s6, s7, s8, s9) => listener(s1.data as T1, s2.data as T2, s3.data as T3,
-          s4.data as T4, s5.data as T5, s6.data as T6, s7.data as T7, s8.data as T8, s9.data as T9),
+      streams: StreamTuple9(pipe1.stream, pipe2.stream, pipe3.stream, pipe4.stream, pipe5.stream, pipe6.stream, pipe7.stream, pipe8.stream, pipe9.stream),
+      initialData: StreamInitialDataTuple9(pipe1.value, pipe2.value, pipe3.value, pipe4.value, pipe5.value, pipe6.value, pipe7.value, pipe8.value, pipe9.value),
+      builder: (context, s1, s2, s3, s4, s5, s6, s7, s8, s9) =>
+          listener(s1.data as T1, s2.data as T2, s3.data as T3, s4.data as T4, s5.data as T5, s6.data as T6, s7.data as T7, s8.data as T8, s9.data as T9),
     );
   }
 }